refactor: Extracted god-component files into several more organized structures, including FarmComponents.kt, AuctionComponents.kt, TradeComponents.kt, GameAssets.kt and GameComponents.kt
